requirements.txt
absl-py==2.1.0
accelerate==0.28.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
black==21.4b2
cachetools==5.3.3
certifi==2023.11.17
click==8.1.7
cloudpickle==3.0.0
cmake==3.20.3
contourpy==1.1.1
cycler==0.12.1
cython==3.0.8
diffdist==0.1
docker-pycreds==0.4.0
einops==0.4.1
executing==2.0.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.1
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
gitpython==3.1.42
google-auth==2.28.1
google-auth-oauthlib==1.0.0
grpcio==1.62.0
h5py==3.7.0
hf-transfer==0.1.6
huggingface-hub==0.20.3
hydra-core==1.3.2
icecream==2.1.2
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
infinibatch==0.1.1
inflect==5.6.0
iopath==0.1.9
jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.1
lazy-loader==0.3
lightning-utilities==0.10.1
lit==17.0.6
markdown==3.5.2
markdown-it-py==3.0.0
markupsafe==2.1.4
matplotlib==3.7.4
mdurl==0.1.2
mmcls==0.25.0
mmcv==2.1.0
mmengine==0.10.3
mmsegmentation==1.2.2
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
natten==0.14.4
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
pathtools==0.1.2
peft==0.4.0
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.8.2
prettytable==3.10.0
promise==2.3
protobuf==3.20.1
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.7
pydeprecate==0.3.2
pydot==2.0.0
pygments==2.17.2
pyparsing==3.1.1
pyproject==1.3.1
pyqt5==5.15.9
pyqt5-plugins==5.15.9.2.3
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytorch-lightning==2.2.1
pytz==2024.1
pywavelets==1.4.1
pyyaml==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rarfile==4.2
regex==2022.6.2
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.8.1
seaborn==0.13.2
sentry-sdk==1.40.6
setproctitle==1.3.3
setuptools==59.5.0
shapely==2.0.3
shortuuid==1.0.12
six==1.16.0
smmap==5.0.1
submitit==1.4.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1-2209072238
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
torch==2.1.1
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.64.0
transformers==4.38.2
triton==2.1.0
typing-extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.15
urllib3==2.1.0
wandb==0.12.20
wcwidth==0.2.13
werkzeug==3.0.1
xformers==0.0.23
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0